requirements.txt
azure-communication-administration==1.0.0b4
azure-communication-sms==1.0.1
azure-core==1.28.0
azure-identity==1.13.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==41.0.1
DateTime==5.1
Flask==2.3.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zope.interface==6.0
